Fix #570 : mise en cache des Pdfs à la validation
on utilise le hook post_valid_callback pour générer le pdf et le placer
en cache
