+++ autonomie/views/task/views.py
@@ -6,6 +6,7 @@
 """
 Base Task views
 """
+import datetime
 import logging
 from pyramid.httpexceptions import HTTPFound
 from autonomie_base.utils.ascii import force_filename
@@ -16,6 +17,7 @@ from autonomie.models.project import (
     Project,
 )
 from autonomie.models.task.task import TaskLine
+from autonomie.models.task import TaskStatus
 from autonomie.forms.tasks.base import (
     get_duplicate_schema,
     get_new_task_schema,
@@ -44,6 +46,7 @@ from autonomie.views import (
     cancel_btn,
     TreeMixin,
 )
+from autonomie.views.status import StatusView
 from autonomie.views.project.routes import (
     PROJECT_ITEM_PHASE_ROUTE,
     PROJECT_ITEM_ROUTE,
@@ -350,17 +353,37 @@ class TaskPdfView(BaseView):
         if self.context.status == 'valid':
             self.context.persist_pdf(filename, pdf_datas)
 
-    def __call__(self):
+    def _get_html_output(self):
+        """
+        Produce an html output of the current context
+        """
+        pdf_css.need()
+        html_string = html(self.request)
+        return html_string
+
+    def _get_filename(self):
+        """
+        Returns the pdf's filename
+        """
         number = self.request.context.internal_number
         label = force_filename(number)
-        filename = u"{0}.pdf".format(label)
+        return u"{0}.pdf".format(label)
 
+    def _get_pdf_buffer(self):
+        """
+        Collect the pdf representation of the current context
+        """
         if self.context.status == 'valid' and self.context.pdf_file is not None:
-            pdf_datas = self.context.pdf_file
+            result = self.context.pdf_file.data_obj
         else:
-            pdf_css.need()
-            html_string = html(self.request)
-            pdf_datas = buffer_pdf(html_string)
+            html_string = self._get_html_output()
+            result = buffer_pdf(html_string)
+        return result
+
+    def __call__(self):
+        filename = self._get_filename()
+        pdf_datas = self._get_pdf_buffer()
+        if self.context.status == 'valid' and self.context.pdf_file is not None:
             self._cache_pdf(filename, pdf_datas)
 
         write_file_to_request(
@@ -483,3 +506,95 @@ class TaskSetDraftView(BaseView):
                 id=self.context.id,
             )
         )
+
+
+class TaskStatusView(StatusView):
+    """
+    View handling base status for tasks (estimation/invoice/cancelinvoice)
+
+    Status related views should implement the validate function to ensure data
+    integrity
+    """
+
+    def validate(self):
+        raise NotImplemented()
+
+    def check_allowed(self, status, params):
+        self.request.context.check_status_allowed(status, self.request)
+
+    def redirect(self):
+        project_id = self.request.context.project.id
+        loc = self.request.route_path(PROJECT_ITEM_ROUTE, id=project_id)
+        if self.request.is_xhr:
+            return dict(redirect=loc)
+        else:
+            return HTTPFound(loc)
+
+    def pre_status_process(self, status, params):
+        if 'comment' in params:
+            self.context.status_comment = params.get('comment')
+            logger.debug(self.context.status_comment)
+
+        if 'change_date' in params and params['change_date'] in ('1', 1):
+            logger.debug("Forcing the document's date !!!")
+            self.context.date = datetime.date.today()
+
+        return StatusView.pre_status_process(self, status, params)
+
+    def pre_wait_process(self, status, params):
+        """
+        Launched before the wait status is set
+
+        :param str status: The new status that should be affected
+        :param dict params: The params that were transmitted by the associated
+        State's callback
+        """
+        self.validate()
+        return {}
+
+    def pre_valid_process(self, status, params):
+        """
+        Launched before the valid status is set
+
+        :param str status: The new status that should be affected
+        :param dict params: The params that were transmitted by the associated
+        State's callback
+        """
+        self.validate()
+        return {}
+
+    def post_valid_process(self, status, params):
+        """
+        Launched after the status is set to valid
+
+        :param str status: The new status that should be affected
+        :param dict params: The params that were transmitted by the associated
+        State's callback
+        """
+        # Generate a cached version of the pdf output
+        view = TaskPdfView(self.context, self.request)
+        filename = view._get_filename()
+        pdf_buffer = view._get_pdf_buffer()
+        self.context.persist_pdf(filename, pdf_buffer)
+
+    def post_status_process(self, status, params):
+        """
+        Launch post status process functions
+
+        :param str status: The new status that should be affected
+        :param dict params: The params that were transmitted by the associated
+        State's callback
+        """
+        logger.debug("post_status_process")
+        logger.debug(self.context.status_comment)
+        # Record a task status change
+        self.context.status_date = datetime.date.today()
+        status_record = TaskStatus(
+            task_id=self.context.id,
+            status_code=status,
+            status_person_id=self.request.user.id,
+            status_comment=self.context.status_comment
+        )
+        self.context.statuses.append(status_record)
+        self.request.dbsession.merge(self.context)
+        StatusView.post_status_process(self, status, params)
